refactor(Portal): move NewsletterSignUpForm to its own file.
